Discovered BUG with the news API, error no. 426
Apparently, the API i had used beforehand has given up and told me to "upgrade" it. i've henceforth learned that it means the current version i have is not paid for, and therefore needs to be "upgraded" in terms of buying a license. i'll do what i can to find an API that works, regardless, you've seen it work mr teacher man.